+++ R/d2b.R
@@ -17,6 +17,7 @@
 #'          function \code{\link{sund2bTooltip}} for more information.
 #' @param breadcrumbs \code{list} of options for customizing the breadcrumb. See the helper
 #'          function \code{\link{sund2bBreadcrumb}} for more information.
+#' @param rootLabel \code{character} to label root node something other than 'root'.
 #' @param height,width  height and width of sunburst htmlwidget containing div
 #'          specified in any valid \code{CSS} size unit.
 #' @param elementId string id as a valid \code{CSS} element id.
@@ -32,6 +33,7 @@ sund2b <- function(
   valueField = "size",
   tooltip = NULL,
   breadcrumbs = NULL,
+  rootLabel = NULL,
   width = NULL, height = NULL, elementId = NULL
 ) {
 
@@ -71,7 +73,8 @@ sund2b <- function(
     ),
     options = list(
       colors = colors,
-      valueField = valueField
+      valueField = valueField,
+      rootLabel = rootLabel
     )
   )